+++ services/cctp-relayer/db/sql/base/message.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+
 	"github.com/ethereum/go-ethereum/common"
 	"gorm.io/gorm/clause"
 
@@ -64,6 +65,7 @@ func (s Store) StoreMessage(ctx context.Context, msg types.Message) error {
 		clauses = clause.OnConflict{
 			Columns: []clause.Column{{Name: MessageHashFieldName}},
 			DoUpdates: clause.AssignmentColumns([]string{
+				DestTxHashFieldName,
 				StateFieldName,
 				NonceFieldName,
 			}),
@@ -72,6 +74,7 @@ func (s Store) StoreMessage(ctx context.Context, msg types.Message) error {
 		clauses = clause.OnConflict{
 			Columns: []clause.Column{{Name: MessageHashFieldName}},
 			DoUpdates: clause.AssignmentColumns([]string{
+				DestTxHashFieldName,
 				StateFieldName,
 			}),
 		}
@@ -116,6 +119,21 @@ func (s Store) GetMessageByOriginHash(ctx context.Context, originHash common.Has
 		First(&message)
 	if dbTx.Error != nil {
 		return nil, fmt.Errorf("failed to get message by hash: %w", dbTx.Error)
+	}
+
+	return &message, nil
+}
+
+// GetMessageByRequestID gets a message by its request id.
+func (s Store) GetMessageByRequestID(ctx context.Context, requestID string) (*types.Message, error) {
+	var message types.Message
+
+	dbTx := s.DB().WithContext(ctx).
+		Model(&types.Message{}).
+		Where(fmt.Sprintf("%s = ?", RequestIDFieldName), requestID).
+		First(&message)
+	if dbTx.Error != nil {
+		return nil, fmt.Errorf("failed to get message by request id: %w", dbTx.Error)
 	}
 
 	return &message, nil